Pull all globals into a context struct.
Addresses GitLab iisue #1 (Empire fails to build from source with GCC
10), treating the actual problem - the C in the ganme was somewhat
archaic in style. With this change there is exactly one global context
that only needs to be declated once and externed once, rather than a
whole bunch of globals declared in one place and externed in another.

The new game_state_t is an untidy pile of disparate etuff. Maybe
it oughtto be split int o a couple of different contexts managed
by different code, but this will do for now and could stand to be
tested before we do anything more elaborate.